+++ Pod/Classes/ViewControllers/ARMapContainerViewController.m
@@ -30,15 +30,15 @@ NSString * const __nonnull SelectedCityNameKey = @"SelectedCityName";
 @end
 
 static UIScrollView *
-FindFirstVerticalScrollView(UIView *view)
+FindFirstScrollView(UIView *view)
 {
     for (UIView *subview in view.subviews) {
-        if ([subview isKindOfClass:UIScrollView.class] && ([(UIScrollView *)subview contentSize].height > subview.frame.size.height)) {
+        if ([subview isKindOfClass:UIScrollView.class]) {
             return (UIScrollView *)subview;
         }
     }
     for (UIView *subview in view.subviews) {
-        UIScrollView *result = FindFirstVerticalScrollView(subview);
+        UIScrollView *result = FindFirstScrollView(subview);
         if (result) return result;
     }
     return nil;
@@ -79,7 +79,7 @@ Since this controller already has to do the above logic, having it handle the Ci
     }];
     [[NSNotificationCenter defaultCenter] addObserverForName:@"ARLocalDiscoveryCityGotScrollView" object:nil queue:nil usingBlock:^(NSNotification * _Nonnull note) {
             if (!wself.rnScrollView) {
-                UIScrollView *foundScrollView = FindFirstVerticalScrollView(wself.cityVC.view);
+                UIScrollView *foundScrollView = FindFirstScrollView(wself.cityVC.view);
                 wself.rnScrollView = foundScrollView;
                 wself.rnScrollView.scrollEnabled = NO;
                 if (foundScrollView) {
@@ -257,14 +257,15 @@ Since this controller already has to do the above logic, having it handle the Ci
 
     BOOL isDrawerOpen = [drawer.drawerPosition isEqualToPosition:PulleyPosition.open];
     [self.cityVC setProperty:@(isDrawerOpen) forKey:@"isDrawerOpen"];
+    self.rnScrollView.scrollEnabled = isDrawerOpen;
+    if (!isDrawerOpen) {
+        [self.rnScrollView setContentOffset:CGPointZero animated:YES];
+    }
 }
 
 - (void)drawerChangedDistanceFromBottomWithDrawer:(PulleyViewController *)drawer distance:(CGFloat)distance bottomSafeArea:(CGFloat)bottomSafeArea
 {
     CGFloat drawerAbovePartialHeight = [drawer partialRevealDrawerHeightWithBottomSafeArea:bottomSafeArea];
-
-    BOOL isDrawerOpen = [drawer.drawerPosition isEqualToPosition:PulleyPosition.open];
-    self.rnScrollView.scrollEnabled = isDrawerOpen;
 
     BOOL shouldHideButtons = distance > drawerAbovePartialHeight;
     if (!self.cityPickerController) {