+++ src/lib/Scenes/City/City.tsx
@@ -1,11 +1,13 @@
-import { Box, color, Flex, Theme } from "@artsy/palette"
+import { color, Flex, Theme } from "@artsy/palette"
+import ScrollableTabBar, { ScrollableTab } from "lib/Components/ScrollableTabBar"
 import { Schema, screenTrack } from "lib/utils/track"
 import React, { Component } from "react"
-import { NativeModules, ScrollView } from "react-native"
+import ScrollableTabView from "react-native-scrollable-tab-view"
+
+import { NativeModules, View } from "react-native"
 import { RelayProp } from "react-relay"
 import styled from "styled-components/native"
-import { BucketKey, BucketResults } from "../Map/bucketCityResults"
-import { FiltersBar } from "../Map/Components/FiltersBar"
+import { BucketResults } from "../Map/bucketCityResults"
 import { EventEmitter } from "../Map/EventEmitter"
 import { MapTab } from "../Map/types"
 import { cityTabs } from "./cityTabs"
@@ -15,18 +17,22 @@ import { EventList } from "./Components/EventList"
 interface Props {
   verticalMargin?: number
   isDrawerOpen?: boolean
+  initialTab?: number
   citySlug: string
   tracking: any
 }
 
 interface State {
   buckets?: BucketResults
-  filter: MapTab
+  filter: MapTab // Used for analytics
   relay: RelayProp
   cityName: string
   citySlug: string
+
+  selectedTab: number
   sponsoredContent: { introText: string; artGuideUrl: string }
 }
+const AllCityMetaTab = 0
 
 const screenSchemaForCurrentTabState = currentSelectedTab => {
   switch (currentSelectedTab) {
@@ -57,11 +63,12 @@ export class CityView extends Component<Props, State> {
     filter: cityTabs[0],
     relay: null,
     cityName: "",
+    selectedTab: AllCityMetaTab,
     citySlug: "",
     sponsoredContent: null,
   }
+
   scrollViewVerticalStart = 0
-  scrollView: ScrollView = null
 
   handleEvent = ({
     filter,
@@ -103,14 +110,15 @@ export class CityView extends Component<Props, State> {
   }
 
   componentDidUpdate() {
-    if (!this.props.isDrawerOpen && this.scrollView) {
-      this.scrollView.scrollTo({ x: 0, y: 0, animated: true })
-    }
-
     if (this.state.buckets) {
       // We have the Relay response; post a notification so that the ARMapContainerViewController can finalize the native UI.
       NativeModules.ARNotificationsManager.postNotificationName("ARLocalDiscoveryQueryResponseReceived", {})
     }
+  }
+
+  setSelectedTab(selectedTab) {
+    this.setState({ selectedTab: selectedTab.i }, this.fireScreenViewAnalytics)
+    EventEmitter.dispatch("filters:change", selectedTab.i)
   }
 
   fireScreenViewAnalytics = () => {
@@ -123,62 +131,63 @@ export class CityView extends Component<Props, State> {
   }
 
   render() {
-    const { buckets, filter, cityName, citySlug } = this.state
+    const { buckets, cityName, citySlug } = this.state
     const { verticalMargin } = this.props
     // bottomInset is used for the ScrollView's contentInset. See the note in ARMapContainerViewController.m for context.
     const bottomInset = this.scrollViewVerticalStart + (verticalMargin || 0)
+
     return (
       buckets && (
         <Theme>
           <>
-            <Box>
+            <Flex style={{ flex: 1 }}>
               <Flex py={1} alignItems="center">
                 <Handle />
               </Flex>
-              <FiltersBar
-                tabs={cityTabs}
-                goToPage={activeIndex => EventEmitter.dispatch("filters:change", activeIndex)}
-              />
-              <ScrollView
-                contentInset={{ bottom: bottomInset }}
-                onLayout={layout => {
-                  this.scrollViewVerticalStart = layout.nativeEvent.layout.y
-                  NativeModules.ARNotificationsManager.postNotificationName("ARLocalDiscoveryCityGotScrollView", {})
-                }}
-                ref={r => {
-                  if (r) {
-                    this.scrollView = r as any
-                  }
+              <ScrollableTabView
+                initialPage={this.props.initialTab || AllCityMetaTab}
+                onChangeTab={selectedTab => this.setSelectedTab(selectedTab)}
+                renderTabBar={props => (
+                  <View>
+                    <ScrollableTabBar {...props} />
+                  </View>
+                )}
+                onLayout={layout => (this.scrollViewVerticalStart = layout.nativeEvent.layout.y)}
+                // These are the ScrollView props for inside the scrollable tab view
+                contentProps={{
+                  contentInset: { bottom: bottomInset },
+                  onLayout: layout => {
+                    this.scrollViewVerticalStart = layout.nativeEvent.layout.y
+                    NativeModules.ARNotificationsManager.postNotificationName("ARLocalDiscoveryCityGotScrollView", {})
+                  },
                 }}
               >
-                {(() => {
-                  switch (filter && filter.id) {
-                    case "all":
-                      return (
-                        <AllEvents
-                          cityName={cityName}
-                          citySlug={citySlug}
-                          key={cityName}
-                          currentBucket={filter.id as BucketKey}
-                          buckets={buckets}
-                          sponsoredContent={this.state.sponsoredContent}
-                          relay={this.state.relay}
-                        />
-                      )
-                    default:
-                      return (
-                        <EventList
-                          key={cityName + filter.id}
-                          bucket={buckets[filter.id]}
-                          type={filter.id as any}
-                          relay={this.state.relay}
-                          cityName={cityName}
-                        />
-                      )
-                  }
-                })()}
-              </ScrollView>
-            </Box>
+                <ScrollableTab tabLabel="All" key="all">
+                  <AllEvents
+                    cityName={cityName}
+                    citySlug={citySlug}
+                    key={cityName}
+                    sponsoredContent={this.state.sponsoredContent}
+                    buckets={buckets}
+                    relay={this.state.relay}
+                  />
+                </ScrollableTab>
+
+                {cityTabs.filter(tab => tab.id !== "all").map(tab => {
+                  return (
+                    <ScrollableTab tabLabel={tab.text} key={tab.id}>
+                      <EventList
+                        key={cityName + tab.id}
+                        bucket={buckets[tab.id]}
+                        type={tab.id}
+                        cityName={cityName}
+                        relay={this.state.relay}
+                      />
+                    </ScrollableTab>
+                  )
+                })}
+              </ScrollableTabView>
+            </Flex>
           </>
         </Theme>
       )