+++ src/lib/Scenes/City/Components/EventList.tsx
@@ -1,8 +1,7 @@
 import { Box, Message, Separator, Serif } from "@artsy/palette"
 import { ShowItemRow } from "lib/Components/Lists/ShowItemRow"
 import Spinner from "lib/Components/Spinner"
-import { BucketKey } from "lib/Scenes/Map/bucketCityResults"
-import { Show } from "lib/Scenes/Map/types"
+import { MapTab, Show } from "lib/Scenes/Map/types"
 import React from "react"
 import { FlatList, NativeScrollEvent, NativeSyntheticEvent } from "react-native"
 import { RelayProp } from "react-relay"
@@ -10,7 +9,7 @@ import { TabFairItemRow } from "./TabFairItemRow"
 
 interface Props {
   bucket: Show[]
-  type: BucketKey | "BMW Art Guide"
+  type: MapTab["id"]
   cityName: string
   header?: string
   relay: RelayProp